# utils/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API CSE-341",
    "description": "Api used for our CSE 341 class",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://cse-341-yf6e.onrender.com",
      "description": "local server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/awesome": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/": {
      "post": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "example": "any"
                  },
                  "lastName": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "all contacts",
        "description": "Used to get all contacts from mongodb",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "create a contact",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Data needed to create a contact",
            "schema": {
              "$ref": "#/components/schemas/contacts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "search by id",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "insert id"
            },
            "description": "Insert user id here to get data"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "summary": "update a contact",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Id of the contact"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Data needed to update a contact",
            "schema": {
              "$ref": "#/components/schemas/updateContacts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "delete a contact",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "contacts": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "<String>"
          },
          "lastName": {
            "type": "string",
            "example": "<String>"
          },
          "email": {
            "type": "string",
            "example": "<String>"
          },
          "favoriteColor": {
            "type": "string",
            "example": "<String>"
          },
          "birthday": {
            "type": "string",
            "example": "<String>"
          }
        }
      },
      "updateContacts": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "<String>"
          },
          "lastName": {
            "type": "string",
            "example": "<String>"
          },
          "email": {
            "type": "string",
            "example": "<String>"
          },
          "favoriteColor": {
            "type": "string",
            "example": "<String>"
          },
          "birthday": {
            "type": "string",
            "example": "<String>"
          }
        }
      }
    }
  },
  "schema": [
    "https"
  ]
}